Renamed RuleParameter to Parameter
