+++ HABApp/parameters/__init__.py
@@ -1,2 +1,2 @@
 
-from .rule_parameter import get_parameter, get_parameter_value, RuleParameter
+from .parameter import Parameter
